+++ lib/animation/fade_in_animation.dart
@@ -27,6 +27,9 @@ part of '../flutter_ex_kit.dart';
 ///   - `FadeInDirection.ttb`: Top to bottom.
 ///   - `FadeInDirection.btt`: Bottom to top.
 class FadeInAni extends StatefulWidget {
+  /// Creates a [FadeInAni] widget.
+  ///
+  /// Requires [child], [delay], [direction], and [fadeOffset] parameters.
   const FadeInAni({
     super.key,
     required this.child,
@@ -35,9 +38,16 @@ class FadeInAni extends StatefulWidget {
     required this.fadeOffset,
   });
 
+  /// The widget to be animated.
   final Widget child;
+
+  /// The duration multiplier for the animation.
   final double delay;
+
+  /// The distance the widget moves during the fade-in animation.
   final double fadeOffset;
+
+  /// The direction from which the widget slides in.
   final FadeInDirection direction;
 
   @override
@@ -45,8 +55,13 @@ class FadeInAni extends StatefulWidget {
 }
 
 class _FadeInAniState extends State<FadeInAni> with TickerProviderStateMixin {
+  /// Controls the animation progress.
   late AnimationController controller;
+
+  /// Controls the opacity transition.
   late Animation<double> opacityAnimation;
+
+  /// Controls the slide transition based on [FadeInDirection].
   late Animation<double> inAnimation;
 
   @override
@@ -60,11 +75,10 @@ class _FadeInAniState extends State<FadeInAni> with TickerProviderStateMixin {
     );
 
     // Define the sliding animation using Tween and animate based on controller.
-    inAnimation =
-        Tween<double>(begin: -widget.fadeOffset, end: 0).animate(controller)
-          ..addListener(() {
-            setState(() {});
-          });
+    inAnimation = Tween<double>(begin: -widget.fadeOffset, end: 0).animate(controller)
+      ..addListener(() {
+        setState(() {});
+      });
 
     // Define the opacity animation, starting from 0 (invisible) and ending at 1 (fully visible).
     opacityAnimation = Tween<double>(begin: 0, end: 1).animate(controller)
@@ -81,14 +95,10 @@ class _FadeInAniState extends State<FadeInAni> with TickerProviderStateMixin {
     // Apply the animation transformations: translate and fade.
     return Transform.translate(
       offset: switch (widget.direction) {
-        FadeInDirection.ltr =>
-          Offset(inAnimation.value, 0), // Move from left to right
-        FadeInDirection.rtl =>
-          Offset(-inAnimation.value, 0), // Move from right to left
-        FadeInDirection.ttb =>
-          Offset(0, inAnimation.value), // Move from top to bottom
-        FadeInDirection.btt =>
-          Offset(0, 0 - inAnimation.value), // Move from bottom to top
+        FadeInDirection.ltr => Offset(inAnimation.value, 0), // Move from left to right
+        FadeInDirection.rtl => Offset(-inAnimation.value, 0), // Move from right to left
+        FadeInDirection.ttb => Offset(0, inAnimation.value), // Move from top to bottom
+        FadeInDirection.btt => Offset(0, -inAnimation.value), // Move from bottom to top
       },
       child: Opacity(
         opacity: opacityAnimation.value, // Apply fade-in effect