+++ lib/enums/fade_in_animation_enum.dart
@@ -1,3 +1,16 @@
 part of '../flutter_ex_kit.dart';
 
-enum FadeInDirection { ttb, btt, ltr, rtl }
+/// Defines the possible directions for the fade-in animation.
+enum FadeInDirection {
+  /// Top to Bottom.
+  ttb,
+
+  /// Bottom to Top.
+  btt,
+
+  /// Left to Right.
+  ltr,
+
+  /// Right to Left.
+  rtl,
+}