+++ lib/flutter_ex_kit.dart
@@ -1,11 +1,7 @@
-library flutter_ex_kit;
-
 import 'dart:developer';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_ex_kit/single_ton.dart';
-
 part 'ui/string_text_extensions.dart';
 part 'ui/scaffold_extensions.dart';
 part 'ui/num_extensions.dart';
@@ -22,8 +18,7 @@ part 'functions/pagination_function.dart';
 part 'functions/page_controller_extensions.dart';
 part 'functions/output.dart';
 
-// part 'responsive/media_query_values.dart';
-part 'responsive/mediq_qury.dart';
+part 'responsive/media_querys.dart';
 part 'theme/color.dart';
 part 'animation/fade_in_animation.dart';
 part 'animation/bottom_animation.dart';